datacollector/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
boto3>=1.18,<2.0
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.20.0
botocore==1.34.127
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
isodate==0.6.1
jmespath==1.0.1
numpy==1.26.4
oauth2client==4.1.3
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDrive==1.3.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Pillow>=8.0.0,<9.0.0
